Arreglo de repositorio

+++ src/main/java/com/Startup/tienda/Repositories/CategoriasRepository.java
@@ -9,7 +9,7 @@ import com.Startup.tienda.Entities.Categorias;
 
 //--Extiende de CrudRepository para poder tener las funciones básicas de un crud
 @Repository
-public interface CategoriasRepository extends JpaRepository<Categorias, Long> {
+public interface CategoriasRepository extends JpaRepository<Categorias, Integer> {
     
     //--Personalizamos el método ya que en CrudRepository es una lista Iterable no una List
     @Override
